Add a general setup for NWListener implementations
